pyproject.toml
[project]
name = "rpg-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = "^3.11"
dependencies = [
    "streamlit (>=1.55.0,<2.0.0)",
    "anthropic (>=0.86.0,<0.87.0)",
    "elevenlabs (>=2.40.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
